Handle node crash and reconnect
If a node gets a error response from another node it will inform all remaining ones and the API for the crash of the latter. This happens via the execute_action and update_alive request. When the crashed node reconnects (first_boot = false) it will inform all the nodes again. The API is updated via a new secure endpoint.